Renamed files, fixed unary minus for coordinates

+++ src/main/java/com/educode/events/communication/MessageReceivedEventBase.java
@@ -8,14 +8,14 @@ import com.educode.nodes.literal.NumberLiteralNode;
  */
 public abstract class MessageReceivedEventBase extends EventTypeBase
 {
-    private final float _channel;
+    private final double _channel;
 
     public MessageReceivedEventBase(NumberLiteralNode numberLiteral)
     {
         this._channel = numberLiteral.getValue();
     }
 
-    public float getChannel()
+    public double getChannel()
     {
         return this._channel;
     }
